+++ internal/manifests/datasource/grafana.go
@@ -1,4 +1,4 @@
-package grafana
+package datasource
 
 import (
 	grafanav1 "github.com/grafana-operator/grafana-operator/v5/api/v1beta1"
